# Conflicts:
#	app/src/main/java/com/example/ken/updish/Activity/UserLoginActivity.java
#	app/src/main/java/com/example/ken/updish/Fragment/PostFragment.java
#	app/src/main/java/com/example/ken/updish/Listener/BottomNagivationListener.java
#	app/src/main/java/com/example/ken/updish/Model/Location.java

+++ app/src/main/java/com/example/ken/updish/BackgroundWorker/PostDetailsBackgroundWorker.java
@@ -160,8 +160,8 @@ public class PostDetailsBackgroundWorker extends AsyncTask<String, Void, String>
             loc.setAddress(obj.getString("address"));
 
             // Ken
-            loc.setLatitude(obj.getLong("latitude"));
-            loc.setLongtitude(obj.getLong("longtitude"));
+            loc.setLatitude(obj.getDouble("latitude"));
+            loc.setLongtitude(obj.getDouble("longtitude"));
 
             tempPost.setLocation(loc);
